Move sous.Config -> cli.Config
